requirements.txt
hikari
ddginternal
python-dotenv
groq
definitely-typed